+++ libs/ruamoko/rua_obj.c
@@ -288,7 +288,7 @@ static const char *
 selector_get_key (void *s, void *_pr)
 {
 	progs_t    *pr = (progs_t *) _pr;
-	return PR_GetString (pr, pr->selector_names[(int) s]);
+	return PR_GetString (pr, pr->selector_names[(long) s]);
 }
 
 static const char *
@@ -404,11 +404,11 @@ static pr_sel_t *
 sel_register_typed_name (progs_t *pr, const char *name, const char *types,
 						 pr_sel_t *sel)
 {
-	int         index;
-	int         is_new = 0;
+	long		index;
+	int			is_new = 0;
 	obj_list   *l;
 
-	index = (int) Hash_Find (pr->selector_hash, name);
+	index = (long) Hash_Find (pr->selector_hash, name);
 	if (index) {
 		for (l = ((obj_list **) pr->selector_sels)[index]; l; l = l->next) {
 			pr_sel_t   *s = l->data;
@@ -444,7 +444,7 @@ sel_register_typed_name (progs_t *pr, const char *name, const char *types,
 	((obj_list **) pr->selector_sels)[index] = l;
 
 	if (is_new)
-		Hash_Add (pr->selector_hash, (void *)index);
+		Hash_Add (pr->selector_hash, (void *) index);
 
 	return sel;
 }
@@ -581,8 +581,8 @@ obj_send_message_in_list (progs_t *pr, pr_method_list_t *method_list,
 		pr_method_t *mth = &method_list->method_list[i];
 		if (mth->method_name && sel_eq (&G_STRUCT (pr, pr_sel_t,
 												   mth->method_name), op)
-			&& !Hash_FindElement (pr->load_methods, (void *)mth->method_imp)) {
-			Hash_AddElement (pr->load_methods, (void *)mth->method_imp);
+			&& !Hash_FindElement (pr->load_methods, (void *) (long) mth->method_imp)) {
+			Hash_AddElement (pr->load_methods, (void *) (long) mth->method_imp);
 
 			PR_ExecuteProgram (pr, mth->method_imp);
 			break;
